pyproject.toml
[tool.poetry]
name = "gakreview"
version = "0.1.0"
description = ""
authors = ["Ian Kim <ruhyun92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.6"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
